# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: Privacy-Focused Email Agent API
  description: REST API for the Privacy-Focused Email Agent
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:5000
    description: Development server
...
paths:
...
  /config:
    get:
      tags:
        - Configuration
      summary: Get application configuration
      description: Retrieves the current application configuration settings
      responses:
        '200':
          description: Configuration settings
          content:
            application/json:
              schema:
                type: object
                properties:
                  Ollama:
                    type: object
                    properties:
                      api_base_url:
                        type: string
                        example: http://localhost:11434
                      model_name:
                        type: string
                        example: llama2
                  App:
                    type: object
                    properties:
                      max_emails_fetch:
                        type: integer
                        example: 50
                  User:
                    type: object
                    properties:
                      signature:
                        type: string
                        example: "Best regards,\nJohn Smith"
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Configuration
      summary: Update application configuration
      description: Updates the application configuration settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                Ollama:
                  type: object
                  properties:
                    api_base_url:
                      type: string
                      example: http://localhost:11434
                    model_name:
                      type: string
                      example: llama2
                App:
                  type: object
                  properties:
                    max_emails_fetch:
                      type: integer
                      example: 50
                User:
                  type: object
                  properties:
                    signature:
                      type: string
                      example: "Best regards,\nJohn Smith"
      responses:
        '200':
          description: Configuration updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Configuration updated successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Invalid request parameters
              code:
                type: integer
                example: 400
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Authentication required
              code:
                type: integer
                example: 401
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: An unexpected error occurred
              code:
                type: integer
                example: 500